requirements.txt
# TCRIE - Temporal Clinical Risk Intelligence Engine
# Author: Nampally Aryan

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0

# Data Validation
pydantic>=2.0.0

# File Upload Support
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0

# Visualization
plotly>=5.15.0

# Testing
pytest>=7.4.0
httpx>=0.24.0
